The CliRel script is now able to train a model, and test data.
Training creates a note for each file, then trains a linearSVD as a model and pickles the model to a specified file.
Testing upickles the model, generates relation labels for the test data and writes the relation to a file in the i2b2 format.

The notes, concepts and relations were abstracted as follows:
A Note is a set of Entries and the base name of the file it came from.
An Entry is a Relation and the tokenized Sentence
A Relation is an ordered pair of Concepts and a label
A Concept is a concept_string, label, lineNo, and start and end indices.

Reading/Writing of Concepts and Relations are now class methods.